Drop start and handler error results.
Dropping it on start makes means plugins are forced to start,
which is both easier to work with and easier to debug, since
they have the chance of reporting any issues on interactions.

Dropping on handlers makes more clear that plugins are supposed
to respond to users and to log, instead of just returning.